Add FromNaN32ps(), converts NaN, preserves signal
FromNaN32ps() converts to binary16 NaN without changing
the quiet bit so it preserves the signalling.
It inlines so it is faster than Fromfloat32().

This was needed because Fromfloat32() is 100% compatible with
AMD and Intel F16C instructions.  Unfortunately, that means
NaN input values are converted to NaN with quiet bit always set
in order to be compatible.  FromNaN32ps() offers an alternative.